Add support for expand
When adding mulitple hosts on a single cloudflare domain, the additional domains will be added as "Subject Alternative Name" and the certificate will need to be updated. Since the default setting in certbot is to ask, this fails.

This change adds support for CERTBOT_EXPAND, which when set to true, will add the `--expand` option on the certbot run.